Move ClrOnlyFact to TestUtilities.Shared
