+++ src/Test/Utilities/Desktop/ConditionalFactAttribute.cs
@@ -7,70 +7,7 @@ using Microsoft.Win32;
 using Xunit;
 
 namespace Roslyn.Test.Utilities
-{
-    public enum ClrOnlyReason
-    {
-        Unknown,
-
-        // The Mono version of ilasm doesn't have all of the features we need to run 
-        // our tests.  In particular it doesn't appear to support the full range of 
-        // modopt operators that our tests invoke.
-        Ilasm,
-
-        // Mono lists certain methods in a different order than the CLR.  For example
-        // Equals, GetHashCode, ToString, etc ... which breaks our tests which hard
-        // code the order. 
-        MemberOrder,
-
-        // Can't emit a PDB.
-        Pdb,
-
-        // The documentation comment compiler has a dependency on a resource in the 
-        // System.Xml assembly.  This is a non-portable / implementation detail 
-        // that Mono doesn't mirror.  We need to make this test more robust so it can
-        // run on all runtimes. 
-        //
-        // See DocumentationCommentCompiler.GetDescription 
-        DocumentationComment,
-
-        // Can't sign. 
-        Signing,
-    }
-
-    public class ClrOnlyFact : FactAttribute
-    {
-        public readonly ClrOnlyReason Reason;
-
-        public ClrOnlyFact(ClrOnlyReason reason = ClrOnlyReason.Unknown)
-        {
-            Reason = reason;
-
-            if (MonoHelpers.IsRunningOnMono())
-            {
-                Skip = GetSkipReason(Reason);
-            }
-        }
-
-        private static string GetSkipReason(ClrOnlyReason reason)
-        {
-            switch (reason)
-            {
-                case ClrOnlyReason.Ilasm:
-                    return "Mono ilasm doesn't support all of the features we need";
-                case ClrOnlyReason.MemberOrder:
-                    return "Mono returns certain symbols in different order than we are expecting";
-                case ClrOnlyReason.Pdb:
-                    return "Can't emit a PDB in this scenario";
-                case ClrOnlyReason.Signing:
-                    return "Can't sign assemblies in this scenario";
-                case ClrOnlyReason.DocumentationComment:
-                    return "Documentation comment compiler can't run this test on Mono";
-                default:
-                    return "Test supported only on CLR";
-            }
-        }
-    }
-
+{   
     public class WindowsOnly : ExecutionCondition
     {
         public override bool ShouldSkip { get { return Path.DirectorySeparatorChar != '\\'; } }
